feat(ui,vm): start to move to a base alert dialog view model
